server/requirements.txt
aiofiles==0.8.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==24.2.0
backoff==2.2.1
boilerpy3==1.0.7
canals==0.2.2
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
construct==2.5.3
dataclasses-json==0.5.14
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
Events==0.5
fastapi==0.115.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.0
h11==0.14.0
haystack==0.42
haystack-ai==2.5.1
haystack-experimental==0.1.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
inflect==7.4.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langsmith==0.1.135
lazy-imports==0.3.1
MarkupSafe==2.1.5
marshmallow==3.22.0
monotonic==1.6
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
num2words==0.5.13
numexpr==2.10.1
numpy==1.26.4
openai==1.47.1
openapi-schema-pydantic==1.2.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.6
posthog==3.6.6
praw==7.7.1
prawcore==2.4.0
prompthub-py==4.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ptrace==0.9.9
pytz==2024.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==0.9.8
requests-toolbelt==1.0.0
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sseclient-py==1.8.0
starlette==0.39.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.13.3
tqdm==4.66.5
transformers==4.30.1
typeguard==4.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.30.6
virtualenv==20.26.4
websocket-client==1.8.0
yarl==1.11.1
